requirements.txt
# Core application dependencies
pandas
pydantic
PyYAML
python-dotenv
pyarrow
click

# Web crawling, data extraction, and LLM/AI
crawl4ai
openai

# Rporting
weasyprint
jinja2

# Testing

